+++ packages/ts-components/src/components/recommended-articles/RecommendedArticles.stories.tsx
@@ -17,7 +17,7 @@ storiesOf('Typescript Component/Recommended Articles', module)
         object: 'RecommendedArticles',
         attrs: {
           event_navigation_action: 'navigation',
-          event_navigation_name: 'widget : relevant article',
+          event_navigation_name: 'widget:relevant article',
           event_navigation_browsing_method: 'click',
           section_details: 'section : <section>',
           article_name: '<headline>',
@@ -33,16 +33,16 @@ storiesOf('Typescript Component/Recommended Articles', module)
   ))
   .add('Recommended Articles - 1 Article', () => (
     <FetchProvider previewData={getArticles(previewData, 1)}>
-      <RecommendedArticles heading="Today's news" />
+      <RecommendedArticles heading="Today's News" />
     </FetchProvider>
   ))
   .add('Recommended Articles - 2 Article', () => (
     <FetchProvider previewData={getArticles(previewData, 2)}>
-      <RecommendedArticles heading="Today's business" />
+      <RecommendedArticles heading="Today's Business" />
     </FetchProvider>
   ))
   .add('Recommended Articles - 3 Article', () => (
     <FetchProvider previewData={previewData}>
-      <RecommendedArticles heading="Today's sport" />
+      <RecommendedArticles heading="Today's Sport" />
     </FetchProvider>
   ));